requirements.txt
# MedicalAgentDiagnosis-MAD Dependencies
# =======================================

# Core Framework
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic-settings>=2.9.0
streamlit>=1.44.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# AI/LLM
openai>=1.60.0
camel-ai[rag,graph-rag]>=0.2.85
google-generativeai>=0.8.0

# Vision Engine (optional - for implementing vision backends)
# torch>=2.0.0
# torchxrayvision>=1.2.0
# scikit-image>=0.24.0
numpy>=1.24.0
Pillow>=10.0.0

# Knowledge Graph
neo4j>=5.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
